+++ src/containerapp/azext_containerapp/commands.py
@@ -124,6 +124,7 @@ def load_command_table(self, _):
         g.custom_command('update', 'update_environment_telemetry_otlp', supports_no_wait=True, exception_handler=ex_handler_factory())
         g.custom_command('remove', 'remove_environment_telemetry_otlp', supports_no_wait=True, confirmation=True, exception_handler=ex_handler_factory())
         g.custom_show_command('show', 'show_environment_telemetry_otlp')
+        g.custom_show_command('list', 'list_environment_telemetry_otlp')
 
     with self.command_group('containerapp github-action') as g:
         g.custom_command('add', 'create_or_update_github_action', exception_handler=ex_handler_factory())